refactor: more deep
