+++ src/main/kotlin/com/solve/domain/admin/contest/controller/AdminContestProblemController.kt
@@ -2,7 +2,7 @@ package com.solve.domain.admin.contest.controller
 
 import com.solve.domain.admin.contest.dto.request.AdminContestProblemAddRequest
 import com.solve.domain.admin.contest.service.AdminContestProblemService
-import com.solve.global.common.BaseResponse
+import com.solve.global.common.dto.BaseResponse
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
 import org.springframework.web.bind.annotation.*
