+++ src/main/kotlin/com/solve/domain/admin/problem/controller/AdminProblemTestCaseController.kt
@@ -3,7 +3,7 @@ package com.solve.domain.admin.problem.controller
 import com.solve.domain.admin.problem.service.AdminProblemTestCaseService
 import com.solve.domain.problem.dto.request.ProblemTestCaseCreateRequest
 import com.solve.domain.problem.dto.request.ProblemTestCaseUpdateRequest
-import com.solve.global.common.BaseResponse
+import com.solve.global.common.dto.BaseResponse
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
 import jakarta.validation.Valid
