+++ src/main/kotlin/com/solve/domain/admin/user/controller/AdminUserConnectionController.kt
@@ -2,7 +2,7 @@ package com.solve.domain.admin.user.controller
 
 import com.solve.domain.admin.user.dto.request.AdminUserConnectionAddRequest
 import com.solve.domain.admin.user.service.AdminUserConnectionService
-import com.solve.global.common.BaseResponse
+import com.solve.global.common.dto.BaseResponse
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
 import org.springframework.web.bind.annotation.*
