+++ src/main/kotlin/com/solve/domain/contest/controller/ContestController.kt
@@ -1,7 +1,7 @@
 package com.solve.domain.contest.controller
 
 import com.solve.domain.contest.service.ContestService
-import com.solve.global.common.BaseResponse
+import com.solve.global.common.dto.BaseResponse
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
 import org.springframework.data.domain.Pageable
