+++ src/main/kotlin/com/solve/domain/problem/controller/ProblemIdeaCommentController.kt
@@ -3,7 +3,7 @@ package com.solve.domain.problem.controller
 import com.solve.domain.problem.dto.request.ProblemIdeaCommentCreateRequest
 import com.solve.domain.problem.dto.request.ProblemIdeaCommentUpdateRequest
 import com.solve.domain.problem.service.ProblemIdeaCommentService
-import com.solve.global.common.BaseResponse
+import com.solve.global.common.dto.BaseResponse
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
 import org.springframework.web.bind.annotation.*
