+++ src/main/kotlin/com/solve/domain/problem/controller/ProblemIdeaLikeController.kt
@@ -1,7 +1,7 @@
 package com.solve.domain.problem.controller
 
 import com.solve.domain.problem.service.ProblemIdeaLikeService
-import com.solve.global.common.BaseResponse
+import com.solve.global.common.dto.BaseResponse
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
 import org.springframework.web.bind.annotation.PathVariable
