+++ src/main/kotlin/com/solve/domain/problem/controller/ProblemSubmitController.kt
@@ -2,7 +2,7 @@ package com.solve.domain.problem.controller
 
 import com.solve.domain.problem.dto.request.ProblemSubmitRequest
 import com.solve.domain.problem.service.ProblemSubmitService
-import com.solve.global.common.BaseResponse
+import com.solve.global.common.dto.BaseResponse
 import io.swagger.v3.oas.annotations.tags.Tag
 import org.springframework.web.bind.annotation.*
