+++ src/main/kotlin/com/solve/domain/problem/domain/entity/ProblemIdeaComment.kt
@@ -1,7 +1,7 @@
 package com.solve.domain.problem.domain.entity
 
 import com.solve.domain.user.domain.entity.User
-import com.solve.global.common.BaseTimeEntity
+import com.solve.global.common.entity.BaseTimeEntity
 import jakarta.persistence.*
 
 @Entity
