+++ src/main/kotlin/com/solve/domain/user/controller/UserMeConnectionController.kt
@@ -2,7 +2,7 @@ package com.solve.domain.user.controller
 
 import com.solve.domain.user.dto.request.UserMeAddConnectionRequest
 import com.solve.domain.user.service.UserMeConnectionService
-import com.solve.global.common.BaseResponse
+import com.solve.global.common.dto.BaseResponse
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
 import org.springframework.web.bind.annotation.*
