+++ src/main/kotlin/com/solve/domain/user/controller/UserMeController.kt
@@ -3,7 +3,7 @@ package com.solve.domain.user.controller
 import com.solve.domain.user.dto.request.UserMePasswordUpdateRequest
 import com.solve.domain.user.dto.request.UserMeUpdateRequest
 import com.solve.domain.user.service.UserMeService
-import com.solve.global.common.BaseResponse
+import com.solve.global.common.dto.BaseResponse
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
 import org.springframework.http.MediaType
