+++ src/main/kotlin/com/solve/domain/user/domain/entity/UserSolved.kt
@@ -1,7 +1,7 @@
 package com.solve.domain.user.domain.entity
 
 import com.solve.domain.problem.domain.entity.Problem
-import com.solve.global.common.BaseTimeEntity
+import com.solve.global.common.entity.BaseTimeEntity
 import jakarta.persistence.*
 import org.hibernate.annotations.OnDelete
 import org.hibernate.annotations.OnDeleteAction
