+++ src/main/kotlin/com/solve/domain/workbook/controller/WorkbookController.kt
@@ -3,7 +3,7 @@ package com.solve.domain.workbook.controller
 import com.solve.domain.workbook.dto.request.CreateWorkbookRequest
 import com.solve.domain.workbook.dto.request.UpdateWorkbookRequest
 import com.solve.domain.workbook.service.WorkbookService
-import com.solve.global.common.BaseResponse
+import com.solve.global.common.dto.BaseResponse
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
 import org.springframework.data.domain.Pageable
