+++ src/main/kotlin/com/solve/domain/workbook/controller/WorkbookLikeController.kt
@@ -1,7 +1,7 @@
 package com.solve.domain.workbook.controller
 
 import com.solve.domain.workbook.service.WorkbookLikeService
-import com.solve.global.common.BaseResponse
+import com.solve.global.common.dto.BaseResponse
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
 import org.springframework.web.bind.annotation.DeleteMapping
