+++ src/main/kotlin/com/solve/domain/workbook/controller/WorkbookProblemController.kt
@@ -2,7 +2,7 @@ package com.solve.domain.workbook.controller
 
 import com.solve.domain.workbook.dto.request.AddWorkbookProblemRequest
 import com.solve.domain.workbook.service.WorkbookProblemService
-import com.solve.global.common.BaseResponse
+import com.solve.global.common.dto.BaseResponse
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
 import org.springframework.web.bind.annotation.*
